requirements.txt
google-cloud-storage
influxdb
ijson
